TablesController fini restent un peu de css

+++ Controllers/TablesController.php
@@ -11,9 +11,36 @@ class TablesController extends Controller
         $table = '\\App\\Models\\'.$tablename;
         $table = new $table();
 
-        if (Form::validate($_POST, ['deleteID'])) {
-            $table->delete($_POST['deleteID']);
-            exit;
+        if (isset($_POST['type'])) {
+            switch ($_POST['type']) {
+                case 'delete':
+                    if (Form::validate($_POST, ['deleteID'])) {
+                        $table->delete($_POST['deleteID']);
+                        exit;
+                    }
+                    break;
+                
+                case 'update':
+                    if (Form::validate($_POST, $table->getChamps())) {
+                        $table->hydrate($_POST);
+                        $table->update();
+                    }
+                    break;
+                
+                case 'create':
+                    $not_needed_champs = [];
+                    foreach ($table::$info_tables as $key => $value) {
+                        if (!empty($value['is_disabled'])) $not_needed_champs[] = $key;
+                    }
+                    if (Form::validate($_POST, array_diff($table->getChamps(), $not_needed_champs))) {
+                        $table->hydrate($_POST);
+                        $idName = 'set' . ucfirst($table->getIdName());
+                        $table->$idName(null);
+                        $table->create();
+                    }
+                    break;
+
+            }
         }
 
         $pageName = 'Table ' . $tablename . ' | Projet BdD';
@@ -23,6 +50,17 @@ class TablesController extends Controller
         for ($i=0; $i < $max; $i++) {
             unset($lines[$i]['password']);
         }
+
+        foreach ($table::$info_tables as $key => $value) {
+            if (strpos($key, 'id') !== false && $value['elementHTML'] == 'select') {
+                $table_2 = '\\App\\Models\\' . ucfirst(substr($key, 3)) . 'Model';
+                $table_2 = new $table_2();
+                $table::$info_tables[$key]['values'] = $table_2->findAll();
+                $table::$info_tables[$key]['name_id'] = $table_2->getIdname();;
+            } 
+        }
+
+        //var_dump($table::$info_tables);
 
         $this->render('/tables/index', compact('pageName', 'tablename', 'table', 'lines'));
     }
